rag-document-qa/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
numpy==1.26.3
openai==1.30.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
google-generativeai==0.7.0
anthropic==0.30.0